move dashboard into component
